requirements.txt
langchain
python-dotenv
langchain-community
langchain-groq
mysql-connector-python
SQLAlchemy
faker